requirements.txt
Flask==2.1.3
imageio==2.20.0
numpy==1.23.1
Pillow==9.2.0
scipy==1.8.1
opencv-python
matplotlib==3.5.2

